stage1: refactor cgroup-related functions
Move them to a new cgroup.go file.
